refactor(frontend): extract pure modules from ChartCard monolith (spec 13, partial)
- extract ChartType union (chart/types.ts), pure helpers (chart/helpers.ts:
  toProperCase/splitLegend/parseDateLoose/useDistinctStrings/useDebounced),
  number/date formatters (chart/format.ts) — verbatim code motion
- ChartCard.tsx 7999 -> 7518 lines; public props + default export unchanged,
  all 6 importers untouched; 0 new tsc errors

Safe partial: renderer/hook registry decomposition (spec steps 3-6) and lib
format dedup deferred to a build-gated follow-up pass (close over local state,
need npm run build to verify).

Spec: v2 Roadmap/P1-frontend/13
